api/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyJWT==2.10.1
requests==2.32.5